rabbit refactoring
