+++ app/Hutch.Relay/Services/Contracts/IDownstreamTaskQueue.cs
@@ -9,15 +9,8 @@ namespace Hutch.Relay.Services.Contracts;
 /// This is quite domain opinionated - not a general purpose queue service;
 /// it cares about Relay SubNodes and Task Types etc.
 /// </summary>
-public interface IRelayTaskQueue
+public interface IDownstreamTaskQueue
 {
-  /// <summary>
-  /// Test if the configured queue backend is available to receive messages, optionally on a specific named queue
-  /// </summary>
-  /// <param name="queueName">Optionally specify a queue name to also ensure is ready</param>
-  /// <returns>True if the queue service is ready to receive items (in the named queue if provided)</returns>
-  public Task<bool> IsReady(string? queueName = null);
-  
   /// <summary>
   /// Send a message with the provided RelayTask Body to the queue
   /// </summary>
@@ -25,7 +18,7 @@ public interface IRelayTaskQueue
   /// <param name="taskBody">The body of the task; may be any valid Task type</param>
   /// <typeparam name="T">The Task type for the provided body</typeparam>
   /// <returns></returns>
-  public Task Send<T>(string subnodeId, T taskBody) where T : TaskApiBaseResponse;
+  public Task Publish<T>(string subnodeId, T taskBody) where T : TaskApiBaseResponse;
 
   /// <summary>
   /// Checks a given SubNode's queue and returns a task if there is one.
